Use `for` loops, not `goto`.
Also use `defer` to make sure `Unlock()` is run before exit.