@itsmontoya 2016-11-09
This pull request is quite the performance hit. At this moment in time, I'll have to deny it. My assumption is most of the slowdown is due to using defer. My suggestion is to break out the for loop change and the defer change as separate pulls. At that point we can do performance testing on each set of changes and see what parts we can implement. 

I do appreciate you taking the time to create a pull request though!

## Before
```
panda@archPanda ~/g/s/g/m/mailbox> go test --bench="Mailbox" --benchtime=3s
BenchmarkMailbox-4              50000000      90.7 ns/op       8 B/op      1 allocs/op
BenchmarkBatchMailbox-4           300000      14243 ns/op      0 B/op      0 allocs/op
```

## After
```
BenchmarkMailbox-4              20000000      190 ns/op        8 B/op      1 allocs/op
BenchmarkBatchMailbox-4           300000      14529 ns/op      0 B/op      0 allocs/op
```